requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pillow==10.1.0
numpy==1.26.2
imagehash==4.3.1
python-multipart==0.0.6
pydantic==2.5.1
pdf2image==1.16.3
poppler-utils==0.1.0
python-jose==3.3.0